backend/app/api/middleware/auth_middleware.py
from fastapi import Request
from jose import jwt, JWTError
from starlette.middleware.base import BaseHTTPMiddleware
from app.core.config import settings
import logging

logger = logging.getLogger(__name__)

PUBLIC_PATHS = {
    "/",
    "/health",
    "/docs",
    "/redoc",
    "/openapi.json",
    f"{settings.API_V1_STR}/auth/login",
    f"{settings.API_V1_STR}/auth/register",
}


class AuthMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        path = request.url.path

        if path.startswith("/ws") or path in PUBLIC_PATHS:
            return await call_next(request)

        # Skip openapi docs sub-paths
        if path.startswith("/docs") or path.startswith("/redoc"):
            return await call_next(request)

        auth_header = request.headers.get("Authorization")
        if auth_header and auth_header.startswith("Bearer "):
            # FIX: use proper slice, not .replace() (safer)
            token = auth_header[len("Bearer "):]
            try:
                payload = jwt.decode(
                    token,
                    settings.SECRET_KEY,
                    algorithms=[settings.ALGORITHM],
                )
                user_id = payload.get("sub")
                if user_id:
                    request.state.user_id = user_id
            except JWTError as e:
                logger.warning(f"JWT validation failed: {e}")

        return await call_next(request)